requirements.txt
nox
numpy
matplotlib

faiss-cpu >= 1.6.3
datasets >= 1.0.1
psutil >= 5.7.0
torch >= 1.4.0
transformers
beautifulsoup4
